refactor: Add GlobalState and GlobalReducers to RootStoreState
This commit adds the GlobalState and GlobalReducers interfaces to the RootStoreState in the feature.d.ts file. The GlobalState interface includes the copy property, which contains the content and show properties for displaying copy in the Copy component. The GlobalReducers interface includes the setCopy case reducer for setting the copy to be displayed. This change is necessary for managing the global state in the application.